kernel_10_26_1_18/jb-mr2-t40-r1-10.26.1.18-20131217-1558-foss/mydroid/external/webkit/LayoutTests/fast/events/touch/touchstart-prevent-default-prevents-click-expected.txt

CONSOLE MESSAGE: line 22: touchstart
PASS touchStartCount is 1
PASS clickCount is 0
PASS successfullyParsed is true

TEST COMPLETE
